requirements.txt
numpy>=1.21.0
torch>=2.0.0
gymnasium>=0.29.0
pygame>=2.5.0

matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0

pandas>=2.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
pillow>=9.5.0